Initial commit. This is an alpha version.
The code should compile a valid WAR file that can be deployed into a Servlet container. However, there has been very little testing and no JUnit test cases exist, yet.

Upcoming releases will include test cases, OAuth integration, and an abstraction of the data layer. Currently, we require MongoDB for the data layer, but this will be changed to be an abstraction of that layer to allow users to implement their own data layers without modifying any other components in the system.