Fixed some bugs and an odd error message.
The error message if the "data" field was missing from a data point that was being uploaded was confusing. This has been updated.

The new AuthFilter does not have access to the parameter list for multipart/form-data POST messages, which means it has no way to get the authentication token from the parameters list. However, the token must be part of the parameter list for POSTs to prevent XSS attacks. Therefore, multipart/form-data POSTs do not work, but normal application/x-www-form-urlencoded requests still work. The only way around this, at this time, is to use a header. This might be acceptable if we use the Authorization header.