Added SQL as a built-in data source.
The data package was refactored to make it easier to add new types of data stores. SQL was added with MySQL-specific SQL code. It would be best if the SQL was based on an ANSI SQL standard, e.g. ANSI-SQL:2008.

Other, smaller changes were made including updating several libraries and extending and better documenting the default configuration file.