Added the Previous and Next headers.
When lists of data are requested, e.g. payload IDs, payload versions, data, two headers were missing, Previous and Next. These should have been URLs that indicated how to get the previous and next parts of the data, if available. If this was the first page or last page of results, the Previous and Next headers would be omitted, respectively.