Added user self-registration.
A user may now make a request to "omh/v1/users/register" with the parameters "username", "password", and "email", and an email will be sent to the user with a link. Once the link is clicked, the user's account will be activated, and they can use the system.

To disable this feature, comment out the @RequestMapping annotation for the registerUser(String, String, HttpServletRequest, HttpServletResponse) and activateUser(String, HttpServletRequest, HttpServletResponse) in the Version1 class. You can comment out the entire function, but, by doing only the @RequestMapping annotation, Spring will fail to connect the functionality to the API, which will result in the same effect.

The email will attempt to use SSL. If the machine that is running this web application does not have access to a SSL certificate for its domain, this can be disabled by removing the "MAIL_PROPERTY_SSL_ENABLED" property from the "MAIL_SESSION_PROPERTIES" in the UserRegistrationRequest class. This can be found near the top of the class in a static block with the other class constants.

References #12