Removed SSL for local SMTP traffic.
The connection from the web application to the SMTP server was defaulting to SSL, however, for most setups, this is unnecessary. The traffic to the SMTP server is local, but, once the local SMTP server begins handling the traffic, it may want to add SSL.

The code is only commented out in case anyone wants to enable SSL because their first hop SMTP server is not local.